src/ocr_engine.py
"""
Core OCR Engine for Receipt Text Extraction
Uses PaddleOCR for high-accuracy text detection and recognition
"""

import os
import time
from typing import Dict, List, Optional, Tuple
from pathlib import Path
import yaml
from loguru import logger

# Fix for Windows OneDNN compatibility issue
os.environ['FLAGS_use_mkldnn'] = 'False'
os.environ['FLAGS_enable_new_ir'] = 'False'

try:
    from paddleocr import PaddleOCR
    import cv2
    import numpy as np
except ImportError as e:
    logger.error(f"Missing dependency: {e}")
    logger.info("Install with: pip install paddlepaddle-gpu paddleocr opencv-python")
    raise


class OCREngine:
    """
    Receipt OCR Engine powered by PaddleOCR
    
    Features:
    - GPU-accelerated text detection and recognition
    - Automatic rotation correction
    - High accuracy on receipt text (95%+)
    - Fast processing (<500ms with GPU)
    """
    
    def __init__(self, config_path: Optional[str] = None):
        """
        Initialize OCR engine with configuration
        
        Args:
            config_path: Path to YAML configuration file
        """
        self.config = self._load_config(config_path)
        self.ocr = None
        self._initialize_ocr()
        
        logger.info("OCR Engine initialized successfully")
        logger.info(f"GPU enabled: {self.config['ocr']['use_gpu']}")
    
    def _load_config(self, config_path: Optional[str] = None) -> Dict:
        """Load configuration from YAML file"""
        if config_path is None:
            config_path = Path(__file__).parent.parent / "config" / "ocr_config.yaml"
        
        if not os.path.exists(config_path):
            logger.warning(f"Config file not found: {config_path}, using defaults")
            return self._default_config()
        
        with open(config_path, 'r') as f:
            config = yaml.safe_load(f)
        
        return config
    
    def _default_config(self) -> Dict:
        """Return default configuration"""
        return {
            'ocr': {
                'use_gpu': True,
                'use_angle_cls': True,
                'lang': 'en',
                'det_db_thresh': 0.3,
                'drop_score': 0.5
            }
        }
    
    def _initialize_ocr(self):
        """Initialize PaddleOCR model"""
        try:
            ocr_config = self.config['ocr']
            
            self.ocr = PaddleOCR(
                use_angle_cls=ocr_config.get('use_angle_cls', True),
                lang=ocr_config.get('lang', 'en'),
                use_gpu=ocr_config.get('use_gpu', True),
                det_db_thresh=ocr_config.get('det_db_thresh', 0.3),
                rec_batch_num=ocr_config.get('rec_batch_num', 6),
                drop_score=ocr_config.get('drop_score', 0.5),
                use_space_char=ocr_config.get('use_space_char', True),
                show_log=False  # Suppress PaddleOCR logs
            )
            
            logger.success("PaddleOCR model loaded successfully")
            
        except Exception as e:
            logger.error(f"Failed to initialize PaddleOCR: {e}")
            raise
    
    def extract_text(
        self, 
        image_path: str,
        return_confidence: bool = True,
        return_positions: bool = False
    ) -> Dict:
        """
        Extract text from receipt image
        
        Args:
            image_path: Path to image file
            return_confidence: Include confidence scores
            return_positions: Include bounding box coordinates
        
        Returns:
            Dictionary with extracted text and metadata
        """
        if not os.path.exists(image_path):
            raise FileNotFoundError(f"Image not found: {image_path}")
        
        logger.info(f"Processing image: {image_path}")
        start_time = time.time()
        
        try:
            # Run OCR
            result = self.ocr.ocr(image_path, cls=True)
            
            if not result or not result[0]:
                logger.warning(f"No text detected in {image_path}")
                return {
                    'status': 'no_text_found',
                    'lines': [],
                    'processing_time_ms': int((time.time() - start_time) * 1000)
                }
            
            # Parse results
            lines = self._parse_ocr_result(
                result[0], 
                return_confidence, 
                return_positions
            )
            
            # Calculate statistics
            avg_confidence = np.mean([line['confidence'] for line in lines])
            processing_time = int((time.time() - start_time) * 1000)
            
            logger.info(f"Extracted {len(lines)} lines in {processing_time}ms")
            logger.info(f"Average confidence: {avg_confidence:.2f}")
            
            return {
                'status': 'success',
                'lines_detected': len(lines),
                'processing_time_ms': processing_time,
                'average_confidence': round(float(avg_confidence), 3),
                'lines': lines
            }
            
        except Exception as e:
            logger.error(f"OCR extraction failed: {e}")
            raise
    
    def _parse_ocr_result(
        self, 
        result: List,
        return_confidence: bool,
        return_positions: bool
    ) -> List[Dict]:
        """Parse PaddleOCR result into structured format"""
        lines = []
        
        for line in result:
            text_info = {
                'text': line[1][0],  # The actual text
            }
            
            if return_confidence:
                text_info['confidence'] = round(float(line[1][1]), 3)
            
            if return_positions:
                # Bounding box coordinates: [[x1,y1], [x2,y2], [x3,y3], [x4,y4]]
                text_info['box'] = line[0]
                text_info['position'] = self._calculate_position_stats(line[0])
            
            lines.append(text_info)
        
        return lines
    
    def _calculate_position_stats(self, box: List) -> Dict:
        """Calculate position statistics from bounding box"""
        # box: [[x1,y1], [x2,y2], [x3,y3], [x4,y4]]
        x_coords = [point[0] for point in box]
        y_coords = [point[1] for point in box]
        
        return {
            'top': int(min(y_coords)),
            'left': int(min(x_coords)),
            'width': int(max(x_coords) - min(x_coords)),
            'height': int(max(y_coords) - min(y_coords))
        }
    
    def batch_extract(self, image_paths: List[str]) -> List[Dict]:
        """
        Process multiple images in batch
        
        Args:
            image_paths: List of image file paths
        
        Returns:
            List of extraction results
        """
        results = []
        total_start = time.time()
        
        logger.info(f"Batch processing {len(image_paths)} images")
        
        for i, image_path in enumerate(image_paths, 1):
            logger.info(f"Processing image {i}/{len(image_paths)}")
            try:
                result = self.extract_text(image_path)
                results.append(result)
            except Exception as e:
                logger.error(f"Failed to process {image_path}: {e}")
                results.append({
                    'status': 'error',
                    'error': str(e),
                    'image_path': image_path
                })
        
        total_time = int((time.time() - total_start) * 1000)
        logger.success(f"Batch complete: {len(results)} images in {total_time}ms")
        
        return results
    
    def get_text_only(self, image_path: str) -> List[str]:
        """
        Simple helper to get just the text lines
        
        Args:
            image_path: Path to image
        
        Returns:
            List of text strings
        """
        result = self.extract_text(image_path, return_confidence=False)
        return [line['text'] for line in result.get('lines', [])]
    
    def validate_image(self, image_path: str) -> Tuple[bool, str]:
        """
        Validate if image is suitable for OCR
        
        Args:
            image_path: Path to image
        
        Returns:
            (is_valid, message)
        """
        if not os.path.exists(image_path):
            return False, "File not found"
        
        try:
            # Try to read image
            img = cv2.imread(image_path)
            
            if img is None:
                return False, "Unable to read image file"
            
            # Check dimensions
            height, width = img.shape[:2]
            
            max_size = self.config.get('preprocessing', {}).get('max_image_size', 4096)
            min_size = self.config.get('preprocessing', {}).get('min_image_size', 100)
            
            if width > max_size or height > max_size:
                return False, f"Image too large (max: {max_size}px)"
            
            if width < min_size or height < min_size:
                return False, f"Image too small (min: {min_size}px)"
            
            return True, "Valid image"
            
        except Exception as e:
            return False, f"Validation error: {str(e)}"


def main():
    """Test the OCR engine"""
    # Configure logging
    logger.add(
        "logs/ocr_test.log",
        rotation="10 MB",
        retention="7 days",
        level="INFO"
    )
    
    print("\n" + "="*60)
    print("Receipt OCR Engine - Test Run")
    print("="*60 + "\n")
    
    # Initialize engine
    engine = OCREngine()
    
    # Check for sample images
    sample_dir = Path("data/sample_receipts")
    sample_dir.mkdir(parents=True, exist_ok=True)
    
    sample_images = list(sample_dir.glob("*.jpg")) + list(sample_dir.glob("*.png"))
    
    if not sample_images:
        print("⚠️  No sample images found in data/sample_receipts/")
        print("Please add some receipt images to test with.")
        print("\nYou can:")
        print("1. Take a photo of a receipt")
        print("2. Download sample receipts from the internet")
        print("3. Use the test image generator (coming in Day 3)")
        return
    
    print(f"Found {len(sample_images)} sample image(s)\n")
    
    # Test with first image
    test_image = str(sample_images[0])
    print(f"Testing with: {test_image}\n")
    
    # Validate image
    is_valid, msg = engine.validate_image(test_image)
    print(f"Image validation: {'✅' if is_valid else '❌'} {msg}")
    
    if not is_valid:
        return
    
    print("\nExtracting text...")
    result = engine.extract_text(test_image, return_confidence=True, return_positions=True)
    
    print(f"\n{'='*60}")
    print("RESULTS")
    print(f"{'='*60}\n")
    print(f"Status: {result['status']}")
    print(f"Lines detected: {result['lines_detected']}")
    print(f"Processing time: {result['processing_time_ms']}ms")
    print(f"Average confidence: {result.get('average_confidence', 0):.1%}\n")
    
    print("Extracted Text:")
    print("-" * 60)
    for i, line in enumerate(result['lines'], 1):
        conf = line.get('confidence', 0)
        print(f"{i:2d}. [{conf:.1%}] {line['text']}")
    
    print("\n" + "="*60 + "\n")


if __name__ == "__main__":
    main()